Always compose QS fragment
This makes it so opening the shade or expanding QS doesn't require
composing all components.

We limit the tiles to listen only when the content should be visible
(and for expanded QS, also when we are not fully collapsed). This
starting to listen causes a small recomposition, but only of the tiles,
not the full grid. Tiles not listening will prevent their UI from
updating.

Notice that other components are not gated by whether we are visible. In
particular:
* Brightness is always composed, which means its viewmodel is always
  collecting. But we don't expect a large issue here.
* Media is explicitly not gated. This guarantees that as soon as a media
  player should be visible, it's composed and measured.

Bug: 389985793
Bug: 399825091
Test: Manual, perfetto shows shorter recomposition on opening shade and
expanding
Test: Manual, dump QSLog and see that tiles are listening at the correct
points
Flag: com.android.systemui.qs_ui_refactor_compose_fragment
Flag: com.android.systemui.always_compose_qs_ui_fragment

Change-Id: I07022bbc0266bf3f1438efcf5759743c9e254f31