[SB][Chips] Fix chip visibility problems with RootMod=on, ChipsMod=off.
We have two separate flows that control the chips: One the controls what
data the chips should show, and another that controls whether the chips
should be visible at all. The current code collects those flows
separately, but that was leading to a problem where each time the chip
data changed, we `launch`ed a new collection of the chip visibility
without cancelling the old collection. This resulted in all the past
chip data being replayed each time the visibility changed.

This CL `combine`s those two separate flows into a single flow and
collects on that flow altogether to avoid this problem.

I think this will fix all the various visibility problems I've seen
over the past week, so optimistically marking them all as fixed here.

Also: Note that this is only relevant when StatusBarRootModernization is
enabled but StatusBarChipsModernization is disabled, and we plan to
launch both flags in the same release. So this code is mostly a stop-gap
if those flags are ever not in sync.

Fixes: 399389764
Fixes: 399372548
Bug: 396471436
Bug: 364653005
Flag: com.android.systemui.status_bar_root_modernization

Test: start then stop screen recording, then open and close shade ->
verify screen recording chip never sticks around (verify with
StatusBarNotifChips both enabled and disabled)
Test: have promoted notification, then open shade, dismiss notification,
and close shade -> verify there's no gap in the status bar between the
clock and the notification icons
Test: have 2 chips then rotate to landscape -> verify first chip
un-squishes and shows text

Test: verify chips with RootModernization disabled work as expected
Test: general smoke test of the status bar, including all chip CUJs

Change-Id: Ib150382be3f0ac0685cf80607092042aece31b64