a11y: Add ScrollPanel
Develop a skeleton scroll panel to manage the autoclick scroll feature.
This panel will appear when the autoclick type is set to 'scroll' and a
pending click is finished. It will be hidden when the autoclick type is
changed from 'scroll' to another type.

Note that this is a temporary UI, Our focus is on getting the scroll
feature working, and we will update the UI to the final
design(screen/W6iBBZJMWWJmmhp) later.

video: http://shortn/_4H94pLSb2s
Design doc: http://b/397391986#comment2

Bug: b/388845721
Test: AutoclickControllerTest
Flag: com.android.server.accessibility.enable_autoclick_indicator
Change-Id: I57b3b375b55872484da12962e0111bf7a8e07fb9