Moved UsbManagerInternal to framework core so other System Services outside of USB can access it (currently there will only be one caller which is UsbDataAdvancedProtectionHook.)
As the UsbManagerInternal is now moved outside of UsbService space, it has been reshaped to be similar to UsbManager to make it more consistent and avoid excess differentiation between Manager clients.

Currently no service is consuming the client. But will be in a subsequent CL.

Bug: 399146898
Change-Id: Iad36ce274a9c444f7c697503a7a4bd39e9df2986
Test: atest UsbServiceTest, manual test
Flag: android.hardware.usb.flags.enable_usb_data_signal_staking_internal