(3/n) Remove usage of isKeyGestureSupported() from KeyGestureController
isKeyGestureSupported is synchronous API that makes system server
dependent on handler and can cause ANR if handlers misbehave.
It was added as a stop gap for multi-key gestures that require
some SysUI info like keyguard state, display state, etc, to decide
if the key needs to be consumed or not at the start of the gesture.

Test: atest KeygestureControllerTests
Bug: 358569822
Bug: 383602794
Flag: EXEMPT refactor
Change-Id: I14abee92ede63e57c4c0d92c3d0eb501bd14a3bd