[Media] Background loading improvements.
Changes the background to a snapshot state. This simplifies the compose
logic so it doesn't need to kick off a coroutine to load the background
and can more easily watch for new background images as they come in.

Also adds a crossfade effect between the unloaded/old and new art as it
comes in.

Bug: 397989775
Test: manually verified in the compose gallery app
Flag: EXEMPT the code is not yet used
Change-Id: Ieacc6e2f2656485fa2dbeee9cad77450a0384546