Add OEM metadata specific permissions to Android Shell
This is required for MediaStore_FilesTest in CtsModernMediaProviderTests. We want to verify update of oem_metadata field in MediaStore using update API but it requires privileged permission which cannot be provided to test app. Hence we need Shell to hold this special permission to verify the API behavior.

Test: n/a
Bug: 352528913
Flag: com.android.providers.media.flags.enable_oem_metadata_update
Change-Id: Iece552da4332bc981eac97461c4705d05bbbe322