Ignore focus change when in RDMV2
In RDMV2, if the device receives a pointer event outside of the
current focused display, the device should not change focus. This
is because the inner display is displaying only a system dialog, and
any interactions with the inner display should not take focus away
from whatever is on the main (outer) display.

In the case of rear camera selfie, changing focus causes the volume
buttons to no longer be sent to the camera app for the shutter action.

Fixes: 398500791
Flag: EXEMPT bugfix
Test: manual test
Test: atest WindowManagerServiceTests
Change-Id: I9eaa1d497d3abb84c1c32079dcb5ad84107a7bc8