Use SystemUIDisplayComponent scope in PerDisplayRepository<CoroutineScope>
This allows classes outside the components to use the scope created inside the component.

As there are several places outside the display component using this scope, this is an intermediate step in the refactor.

Future classes requiring the coroutine scope are expected to be created inside the display subcomponent

Bug: 403481022
Test: DisplayScopeRepositoryInstanceProviderTest, DisplayComponentPerDisplayRepositoryTest
Flag: NONE - refactor without behaviour changes
Change-Id: I95cf815cbfdc3c334ff404434d7ff847a3342330