[flexiglass] SUW visible again.
The setup wizard (SUW) wasn't shown because Flexiglass was incorrectly
reporting to WindowManager (WM) that the keyguard is visible.

By changing the lockscreenVisibilityWithScenes flow to take
SceneInteractor.isVisible into account, we fix this issue.

The reason this works is because, when Flexiglass is on,
WindowManagerLockscreenVisibilityInteractor.surfaceBehindVisibility is
what determines what System UI reports to WM and surfaceBehindVisibility
depends on lockscreenVisibilityWithScenes. The SceneInteractor.isVisible
state is fed by the SceneContainerStartable, setting it to false when
device provisioning is still ongoing (which is the state when SUW is
showing).

Fix: 396690884
Test: unit tests added
Test: manually verified that the SysUiState dump ($ adb shell dumpsys activity service com.android.systemui/.SystemUIService | grep SysUiState -A 10)
is the same when flexiglass is on as it is when it's off
Test: manually verified that SUW is visible when Flexiglass is on and
that skipping SUW makes the scene container be visible again
Flag: com.android.systemui.scene_container

Change-Id: I01d8e19250346bfe02ef730b44b41a2dbb63afd0