[Flexiglass] Use disableSwipesWhenScrolling for nested scrolling as well
Modifier.disableSwipesWhenScrolling disables the STL user actions if
anything below it in the Compose tree has consumed any non-zero scroll
amount during the current gesture. This is important, because a swipe up
can close the shade, and we don't want that to happen while the stack is
being scrolled.

This CL changes the orders of modifiers on the scrollable stack, so we
do this behavior for both nested and regular scrolling.

Bug: 394813692
Test: verify that a nested scroll doesn't close the shade
       - Open Single shade
       - Have enough notifications to scroll, BUT
       - Have a small enough scroll range that only needs nested scrolling
       - Scroll to the bottom with one long continuous motion
       - CHECK: the shade stays open
Flag: com.android.systemui.scene_container
Change-Id: I7624d9326192185a0d7e6a51616fc846892bbf26