[PiP on Desktop] Fix multi-activity PiP expand
Recall: http://recall/clips/01dd5aa4-2628-45a7-852e-652faeb5bef0

With multi-activity PiP (e.g. Netflix), the PiP task is reparented to
the parent task when PiP is expanded. This means that the PiP will be
restored to the parent's windowing mode regardless of what windowing
mode we set the PiP task to. This led to mismatched windowing mode in
the following cases due to the parent retaining the pre-PiP windowing
mode:
- PiP from fullscreen, enter Desktop Windowing and expand PiP -> PiP
  restores to fullscreen
- PiP from freeform, exit Desktop Windowing and expand PiP -> PiP
  restores to freeform

This CL adds a change to the parent task's windowing mode and bounds to
the expand PiP transition based on whether Desktop session is active at
the time of the transition.

Bug: 377707798
Test: atest DesktopPipTransitionControllerTest PipDesktopStateTest PipSchedulerTest
Flag: com.android.window.flags.enable_desktop_windowing_pip

Change-Id: Ie0bf6656bbf5985aab0e6d4e4f748ff81bb8af8d