Change Image used in SmallTileContent
This allows using ContentScale.Crop without clipping. So the image is
scaled so the shortest size matches the desired size, and centered. But
animation can get out of the bounding box (in one of the directions).

Because SmallTileContent is always located inside an element that clips,
the icon will not escape its container.

Also, use a lambda to pass the ColorFilter to help with animation.

Test: manual, icons look the same
Bug: 388292225
Flag: com.android.systemui.qs_ui_refactor_compose_fragment
Flag: com.android.systemui.icon_refresh_2025

Change-Id: I3ed21d5aa2e41e9dc96c3709dcd138353a35febf