Show launcher icons behind dragged task
Change the order of layers behind a dragged task when desktop is not
available.
In this case drag can lead to split or bubbles. And in both of these
cases the task will launch on top of home screen.
Make sure that the entire home screen, including launcher icons, is
visible behind the dragged task when drag can't take the user to
desktop.

Bug: 394651724
Test: atest DragToDesktopTransitionHandlerTest
Flag: com.android.wm.shell.enable_bubble_to_fullscreen
Change-Id: Ic72547f50e7273516a3587ef9b8faadf064739ba