Reset imeShowing when not serverVisible
Previously we would only reset the imeShowing state of the
ImeInsetsSourceProvider if the server visibility changed during the
onPostLayout call. However, this can already be set to false from a
previous call, as is the case when the windowContainer is removed, and
set to null. In this case we wouldn't be serverVisible, but imeShowing
would have a stale true value.

This fixes the issue by replacing the wasServerVisible check with
verifying the current value of imeShowing.

Flag: EXEMPT bugfix
Bug: 406736702
Test: ImeInsetsSourceProvider#testOnPostLayout_resetImeShowingWhenAlreadyNotServerVisible
Change-Id: I5d3745e292d101c0f7a20cbc6a0a10adf6f89672