Add UPDATE_APP_OPS_STATS to com.android.server.telecom.
As part of our mainline efforts we are migrating away from hidden APIs.
We used to call PermissionChecker which is a hidden API.  We are now using
the equivalent PermissionManager APIs which require the caller to have
android.permission.UPDATE_APP_OPS_STATS permission.

We need this so that we can check if an InCallService holds the
android.permission.MANAGE_ONGOING_CALLS. This is an app-op permission that
Companion Device Manager grants/revokes to an InCallService in a
watch companion app based on if a watch association is present and
active.

Bug: 308011328
Test: Run CTS tests with flag on to validate app-ops check behavior.
Flag: com.android.server.telecom.flags.resolve_hidden_dependencies_two
Change-Id: I20c13ecbf4e7da924fc388b7e7e12a51a78184f6