Cache settings per deviceId in ActivityThread
With go/device-aware-settings , we can have custom
settings values for every deviceId. CoreSettingsObserver
in ActivityManagerService fetches settings for all devices
and sends them to ActivityThread to cache in the app process.
This is required as an application can be running on multiple
devices at the same time (eg: one activity on phone and
another on Chromebook).

Test: atest CoreSettingsObserverTest
Test: atest ActivityThreadTest
Test: atest ActivityThreadClientTest
Bug: 371801645
Flag: android.companion.virtualdevice.flags.device_aware_settings_override
Change-Id: I872b39ef944073a4169c59256cefe35a95397941