Remove no-op calls to ActivityManager
When a FGS notification is first added and the session is already
considered "engaged", there is no need to tell the ActivityManager
about it as it will consider new FGSs as active anyway.

When a notification is removed, the associated FGS must be stopped
anyway, so no need to tell the ActivityManager whether the session
is considered user-engaged.

Flag: com.android.media.flags.enable_notifying_activity_manager_with_media_session_status_change
Test: ActivityManagerNotifyMediaFGSTypeTest covers use cases, should not
be affected
Bug: 281762171

Change-Id: Ie9c3fe191e22856ed763bc999f339ef8e5351145